+++ dll/ace/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.22](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@2.0.21...@dll/ace@2.0.22) (2020-10-22)
+
+**Note:** Version bump only for package @dll/ace
+
+## [2.0.21](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@2.0.20...@dll/ace@2.0.21) (2020-10-20)
+
+**Note:** Version bump only for package @dll/ace
+
+## [2.0.20](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@2.0.19...@dll/ace@2.0.20) (2020-10-19)
+
+**Note:** Version bump only for package @dll/ace
+
+## [2.0.19](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@2.0.18...@dll/ace@2.0.19) (2020-10-16)
+
+**Note:** Version bump only for package @dll/ace
+
 ## [2.0.18](https://git.easyops.local/anyclouds/next-core/compare/@dll/ace@2.0.17...@dll/ace@2.0.18) (2020-10-15)
 
 **Note:** Version bump only for package @dll/ace