+++ packages/brick-dll/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.2.3...@easyops/brick-dll@2.2.4) (2020-10-22)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [2.2.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.2.2...@easyops/brick-dll@2.2.3) (2020-10-20)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [2.2.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.2.1...@easyops/brick-dll@2.2.2) (2020-10-19)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [2.2.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.2.0...@easyops/brick-dll@2.2.1) (2020-10-16)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
 # [2.2.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.1.5...@easyops/brick-dll@2.2.0) (2020-10-15)
 
 ### Features