+++ packages/brick-kit/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.11.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-kit@2.10.0...@easyops/brick-kit@2.11.0) (2020-10-22)
+
+### Features
+
+- api analytics ([45b1967](https://git.easyops.local/anyclouds/next-core/commits/45b1967))
+
+# [2.10.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-kit@2.9.2...@easyops/brick-kit@2.10.0) (2020-10-20)
+
+### Features
+
+- getBasePath ([e3e3901](https://git.easyops.local/anyclouds/next-core/commits/e3e3901))
+
+## [2.9.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-kit@2.9.1...@easyops/brick-kit@2.9.2) (2020-10-19)
+
+**Note:** Version bump only for package @easyops/brick-kit
+
+## [2.9.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-kit@2.9.0...@easyops/brick-kit@2.9.1) (2020-10-16)
+
+**Note:** Version bump only for package @easyops/brick-kit
+
 # [2.9.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-kit@2.8.1...@easyops/brick-kit@2.9.0) (2020-10-15)
 
 ### Features