+++ packages/next-docs/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.6](https://git.easyops.local/anyclouds/next-core/compare/@easyops/next-docs@1.0.5...@easyops/next-docs@1.0.6) (2020-10-22)
+
+**Note:** Version bump only for package @easyops/next-docs
+
+## [1.0.5](https://git.easyops.local/anyclouds/next-core/compare/@easyops/next-docs@1.0.4...@easyops/next-docs@1.0.5) (2020-10-20)
+
+**Note:** Version bump only for package @easyops/next-docs
+
+## [1.0.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/next-docs@1.0.3...@easyops/next-docs@1.0.4) (2020-10-19)
+
+**Note:** Version bump only for package @easyops/next-docs
+
+## [1.0.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/next-docs@1.0.2...@easyops/next-docs@1.0.3) (2020-10-16)
+
+**Note:** Version bump only for package @easyops/next-docs
+
 ## [1.0.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/next-docs@1.0.1...@easyops/next-docs@1.0.2) (2020-10-15)
 
 **Note:** Version bump only for package @easyops/next-docs