+++ packages/webpack-config-factory/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.0.3...@easyops/webpack-config-factory@2.0.4) (2020-10-19)
+
+### Bug Fixes
+
+- downgrade postcss to v7 ([cdbcf32](https://git.easyops.local/anyclouds/next-core/commits/cdbcf32))
+- revert postcss-\* and css-loader since postcss v8 is not widely supported ([f24f409](https://git.easyops.local/anyclouds/next-core/commits/f24f409))
+
+## [2.0.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.0.2...@easyops/webpack-config-factory@2.0.3) (2020-10-16)
+
+**Note:** Version bump only for package @easyops/webpack-config-factory
+
 ## [2.0.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/webpack-config-factory@2.0.1...@easyops/webpack-config-factory@2.0.2) (2020-10-15)
 
 ### Bug Fixes